+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Allow configuration via environment variables instead of CLI parameters for all supported options
 - Reenable e2e tests and enhance REST API ([#3836](https://github.com/hoprnet/hoprnet/pull/3836))
 - Fix decoding error when API token contains certain characters
+- Automatically resend queuing transactions when provider is reset or a node receives sufficient native tokens
 - Make environment variables for `hoprd` accessible in Avado package configuration ([#3885](https://github.com/hoprnet/hoprnet/pull/3885))
 - Add Docker image (hopr-hardhat) running hardhat using the Hopr environment and smart contracts (useful for testing and development)
 - Add Docker image (hopr-pluto) running a full hoprd test cluster for local dApp testing